- Correction d'un bug lors de la suppression (suppression en base mais erreur sur l'affichage). - Lors de la mise à jour d'une case, la recherche ne prenait pas en compte la nouvelle valeur

+++ DataTable/php/Table.php
@@ -65,14 +65,16 @@ class Table {
 		if (is_dir($directoryPath)) {
 			if ($directory = opendir($directoryPath)) {
 				while (($file = readdir($directory)) !== false) {
-					$filePath = $directoryPath . $file;
-					$update = filemtime($filePath);
-					$date = time();
-					$dif = $date - $update;
-					
-					// Une heure en seconde => 60 * 60 = 3600
-					if ($dif > 3600)
-						unlink($filePath);
+					if ($file != '.' && $file != '..') {
+						$filePath = $directoryPath . $file;
+						$update = filemtime($filePath);
+						$date = time();
+						$dif = $date - $update;
+							
+						// Une heure en seconde => 60 * 60 = 3600
+						if ($dif > 3600 && file_exists($filePath))
+							unlink($filePath);
+					}
 				}
 				closedir($directory);
 			}
